!!! Very dangerous fix in sizeof(ts)

+++ core-lib/rtp/RTPOpusDepayProcessor.cpp
@@ -10,7 +10,7 @@ RTPOpusDepayProcessor::RTPOpusDepayProcessor()
 : m_maxWidth(0)
 , m_maxHeight(0)
 , m_frameSize(0)
-, m_frameTimestamp(std::numeric_limits<std::uint32_t>::max())
+, m_frameTimestamp(std::numeric_limits<std::uint64_t>::max())
 , m_rtpHelper(new RTPHelper)
 {
 }
@@ -41,7 +41,7 @@ void RTPOpusDepayProcessor::Process(const udp_packet_ptr& pkt) {
         auto packetTimestamp = this->m_rtpHelper->timestamp();
 
         // we got packet at the first time
-        if (m_frameTimestamp == std::numeric_limits<std::uint32_t>::max())
+        if (m_frameTimestamp == std::numeric_limits<std::uint64_t>::max())
         {
             m_frameTimestamp = packetTimestamp;
         }
